// Pollr.AdminUI/ClientApp/src/app/shared/models/answer.model.ts
/*---------------------------------------------------------------------------------------------
 *  Copyright (c) John Duckmanton.
 *  All rights reserved.
 *  Licensed under the MIT License. See LICENSE in the project root for license information.
 *--------------------------------------------------------------------------------------------*/
export class Answer {
  id: number = 0;
  answerText: string;
  imagePath: string;
  voteCount: number;
  isDisabled: boolean;
}